requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
Django==4.2.2
django-autocomplete-light==3.9.7
django-bootstrap-datepicker-plus==5.0.4
django-crispy-forms==2.1
djangorestframework==3.14.0
Pillow==10.1.0
pydantic==2.1.1
pydantic_core==2.4.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
